+++ server/src/organization/organization.service.ts
@@ -12,6 +12,10 @@ import { ClientService } from '../client/client.service';
 import { v4 } from 'uuid';
 import { PrismaService } from '../prisma/prisma.service';
 import { OrganizationDto, UpdateOrganizationDataDto } from './organization.dto';
+import { AppAbility, CaslAbilityFactory } from '../casl/casl-ability.factory';
+import { Action } from '../casl/action.enum';
+import { subject } from '@casl/ability';
+import { accessibleBy } from '@casl/prisma';
 
 export const defaultEventData = {
   name: 'Default Event',
@@ -19,7 +23,7 @@ export const defaultEventData = {
   requiredMembers: 1,
   difficulty: DifficultyMode.NORMAL,
   timeLimitation: TimeLimitationType.PERPETUAL,
-  indexable: false,
+  indexable: true,
   endTime: new Date('2060'),
   latitude: 42.44755580740012,
   longitude: -76.48504614830019,
@@ -44,6 +48,7 @@ export class OrganizationService {
   constructor(
     private prisma: PrismaService,
     private clientService: ClientService,
+    private abilityFactory: CaslAbilityFactory,
   ) {}
 
   async makeDefaultEvent() {
@@ -139,41 +144,25 @@ export class OrganizationService {
   async emitUpdateOrganizationData(
     organization: Organization,
     deleted: boolean,
-    admin?: boolean,
-    user?: User,
+    target?: User,
   ) {
     const dto: UpdateOrganizationDataDto = {
       organization: deleted
-        ? organization.id
+        ? { id: organization.id }
         : await this.dtoForOrganization(organization),
       deleted,
     };
 
-    // Only admin data for now
-    if (user) {
-      await this.clientService.sendUpdate(
-        'updateOrganizationData',
-        user.id,
-        true,
-        dto,
-      );
-    } else {
-      await this.clientService.sendUpdate(
-        'updateOrganizationData',
-        organization.id,
-        true,
-        dto,
-      );
-    }
-  }
-
-  async isManagerOf(
-    org: Organization | { id: string },
-    user: User | { id: string },
-  ) {
-    return !!(await this.prisma.organization.findFirst({
-      where: { id: org.id, managers: { some: { id: user.id } } },
-    }));
+    await this.clientService.sendProtected(
+      'updateOrganizationData',
+      target?.id ?? organization.id,
+      dto,
+      {
+        id: organization.id,
+        subject: subject('Organization', organization),
+        dtoField: 'organization',
+      },
+    );
   }
 
   private genAccessCode() {
@@ -181,34 +170,58 @@ export class OrganizationService {
   }
 
   /** Update/insert a organization group */
-  async upsertOrganizationFromDto(organization: OrganizationDto) {
-    const exists =
-      (await this.prisma.organization.count({
-        where: { id: organization.id },
-      })) > 0;
-
-    return await this.prisma.organization.upsert({
+  async upsertOrganizationFromDto(
+    ability: AppAbility,
+    organization: OrganizationDto,
+  ) {
+    let org = await this.prisma.organization.findFirst({
       where: { id: organization.id },
-      create: {
-        name: organization.name,
-        accessCode: Math.floor(Math.random() * 0xffffff).toString(16),
-        members: {
-          connect: organization.members.map(id => ({ id })),
-        },
-        events: {
-          connect: organization.events.map(id => ({ id })),
-        },
-        specialUsage: 'NONE',
-      },
-      update: {
-        members: {
-          set: organization.members.map(id => ({ id })),
-        },
-        events: {
-          set: organization.events.map(id => ({ id })),
-        },
-      },
     });
+
+    const assignData = {
+      name: organization.name,
+      members: {
+        connect: organization.members?.map(id => ({ id })),
+      },
+      events: {
+        connect: organization.events?.map(id => ({ id })),
+      },
+      specialUsage: OrganizationSpecialUsage.NONE,
+    };
+
+    if (
+      org &&
+      (await this.prisma.organization.findFirst({
+        select: { id: true },
+        where: { AND: [accessibleBy(ability, Action.Update).Organization] },
+      }))
+    ) {
+      const updateData = await this.abilityFactory.filterInaccessible(
+        assignData,
+        subject('Organization', org),
+        ability,
+        Action.Update,
+      );
+
+      org = await this.prisma.organization.update({
+        where: { id: org.id },
+        data: updateData,
+      });
+    } else if (!org && ability.can(Action.Create, 'Organization')) {
+      const data = {
+        ...assignData,
+        name: assignData.name ?? 'New organization',
+        accessCode: Math.floor(Math.random() * 0xffffff).toString(16),
+      };
+
+      org = await this.prisma.organization.create({
+        data,
+      });
+
+      console.log(`Created organization ${org.id}`);
+    }
+
+    return org;
   }
 
   async addAllAdmins(org: Organization) {
@@ -217,14 +230,24 @@ export class OrganizationService {
     });
 
     for (const user of users) {
-      this.clientService.subscribe(user, org.id, true);
+      this.clientService.subscribe(user, org.id);
     }
   }
 
-  async removeOrganization(id: string) {
-    await this.prisma.organization.delete({
-      where: { id },
-    });
+  async removeOrganization(ability: AppAbility, id: string) {
+    if (
+      await this.prisma.organization.findFirst({
+        where: {
+          AND: [{ id }, accessibleBy(ability, Action.Delete).Organization],
+        },
+      })
+    ) {
+      await this.prisma.organization.delete({
+        where: { id },
+      });
+
+      console.log(`Deleted organization ${id}`);
+    }
   }
 
   async ensureFullAccessIfNeeded(potentialAdmin: User) {
@@ -239,6 +262,7 @@ export class OrganizationService {
           where: { id: potentialAdmin.id },
           data: {
             managerOf: { connect: { id } },
+            memberOf: { connect: { id } },
           },
         });
       }
@@ -246,35 +270,50 @@ export class OrganizationService {
   }
 
   async addManager(
-    manager: User,
+    ability: AppAbility,
     potentialManagerEmail: string,
     organizationId: string,
   ) {
-    const org = await this.prisma.organization.findFirstOrThrow({
-      where: { id: organizationId },
+    const org = await this.prisma.organization.findFirst({
+      where: {
+        AND: [
+          accessibleBy(ability, Action.Manage).Organization,
+          { id: organizationId },
+        ],
+      },
     });
 
-    if ((await this.isManagerOf(manager, org)) || manager.administrator) {
-      const potentialManager = await this.prisma.user.findFirstOrThrow({
-        where: { email: potentialManagerEmail },
-      });
-      await this.prisma.organization.update({
-        where: { id: org.id },
-        data: { managers: { connect: { id: potentialManager.id } } },
-      });
+    const potentialManager = await this.prisma.user.findFirst({
+      where: { email: potentialManagerEmail },
+    });
 
-      await this.prisma.user.update({
-        where: { id: potentialManager.id },
-        data: { managerOf: { connect: { id: org.id } } },
-      });
-      console.log('Manager Added');
+    if (!potentialManager || !org) {
+      return false;
     }
+
+    await this.prisma.organization.update({
+      where: { id: org.id },
+      data: {
+        managers: { connect: { id: potentialManager.id } },
+        members: { connect: { id: potentialManager.id } },
+      },
+    });
+
+    await this.prisma.user.update({
+      where: { id: potentialManager.id },
+      data: { managerOf: { connect: { id: org.id } } },
+    });
+
+    console.log(`Manager ${potentialManagerEmail} Added`);
+
+    return true;
   }
 
   async joinOrganization(user: User, code: string) {
     const org = await this.prisma.organization.findFirstOrThrow({
       where: { accessCode: code },
     });
+
     await this.prisma.organization.update({
       where: { id: org.id },
       data: { members: { connect: { id: user.id } } },
